+++ resolwe/flow/elastic_indexes/__init__.py
@@ -1,3 +1,5 @@
 """Elastic search indexes for Resolwe models."""
 
+from .collection import CollectionDocument, CollectionIndex
 from .data import DataDocument, DataIndex
+from .entity import EntityDocument, EntityIndex